connected User table and Task table

+++ app/routers/tasks.py
@@ -1,10 +1,11 @@
 from typing import Annotated
 from fastapi import APIRouter, Path, HTTPException, Depends
+from sqlmodel import select
 
 from ..models import Task, User
 from ..db_engine import SessionDep
 from ..dependencies import get_request_user
-from ..permissions import check_user_ownership
+from ..permissions import check_task_ownership
 
 task_router = APIRouter(
     prefix="/tasks"
@@ -12,7 +13,7 @@ task_router = APIRouter(
 
 @task_router.get("/")
 async def task_list(session: SessionDep, request_user: User = Depends(get_request_user)):
-    tasks = session.query(Task).all()
+    tasks = session.exec(select(Task).where(Task.user_id == request_user.id)).all()
     return tasks
 
 @task_router.get("/{task_id}")
@@ -22,17 +23,20 @@ async def task_detail(
         request_user: User = Depends(get_request_user)
 ):
     task = session.get(Task, task_id)
+    check_task_ownership(task, request_user)
+
     if not task:
         return HTTPException(status_code=404, detail="Task not found")
     else:
         return task
 
-@task_router.post("/")
+@task_router.post("/", response_model=Task)
 async def create_task(
         task: Task,
         session: SessionDep,
         request_user: User = Depends(get_request_user)
 ):
+    task.user_id = request_user.id
     session.add(task)
     session.commit()
     session.refresh(task)
@@ -45,17 +49,24 @@ async def update_task(
         session: SessionDep,
         request_user: User = Depends(get_request_user)
 ):
+    # Check if user can change user_id in Task instance
+
     task_db = session.get(Task, task_id)
+    check_task_ownership(task_db, request_user)
 
     if not task:
-        return HTTPException(status_code=404, detail="Task not found")
-    else:
-        task_data = task.model_dump(exclude_unset=True)
-        task_db.sqlmodel_update(task_data)
-        session.add(task_db)
-        session.commit()
-        session.refresh(task_db)
-        return task
+        raise HTTPException(status_code=404, detail="Task not found")
+
+    task_data = task.model_dump(exclude_unset=True)
+
+    if not task_data.get("user_id") == request_user.id:
+        raise HTTPException(status_code=403, detail="Task owner cannot be changed")
+
+    task_db.sqlmodel_update(task_data)
+    session.add(task_db)
+    session.commit()
+    session.refresh(task_db)
+    return task
 
 @task_router.delete("/{task_id}")
 async def delete_task(
@@ -64,6 +75,7 @@ async def delete_task(
         request_user: User = Depends(get_request_user)
 ):
     task = session.get(Task, task_id)
+    check_task_ownership(task, request_user)
 
     if task:
         session.delete(task)
